title screen working, draw clickhandlers created, and input CLEAR!
