+++ utils/color-utils.js
@@ -1,0 +1,5 @@
+function randomColor() {
+    return [Math.random() * 255, Math.random() * 255, Math.random() * 255]
+};
+
+module.exports = { randomColor };
